Fix for code scanning alert: Multiplication result converted to larger type

+++ src/RageSurfaceUtils.cpp
@@ -493,7 +493,7 @@ static bool blit_same_type( const RageSurface *src_surf, const RageSurface *dst_
 	// If possible, memcpy the whole thing.
 	if( src_surf->w == width && dst_surf->w == width && src_surf->pitch == dst_surf->pitch )
 	{
-		memcpy( dst, src, height*src_surf->pitch );
+		memcpy( dst, src, static_cast<size_t>(height) * src_surf->pitch );
 		return true;
 	}
